Adds documentation on the frames for asset data (#742)
# Description

This MR adds docstrings in articulation and rigid object data to clarify
the exact frames they are reported in. This was previously causing
confusion as PhysX reports the transforms in actor frame and velocities
in CoM frame. Right now, there is no mechanism to get all this data in
the same frame, so the best we can do is document behavior.

## Type of change

- This change requires a documentation update

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there