Fixes location of images for task workflows (#1349)
This MR fixes the regression in #1337. The images were located in the
wrong directory. Please be careful.

P.S.: The images are very low in resolution, and they don't support the
dark theme. It would be great to include this as part of this MR to
avoid this issue.

- Bug fix (non-breaking change which fixes an issue)

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there