Adds missing "not" in  error message in setup_vscode.py (#1615)
# Description

Adds a missing word in an error message.
Also fixes a broken link in the pull request template.

(I did not enter an issue since it seems like a tiny change, but will if
you'd like me to.)

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [ ] I have made corresponding changes to the documentation - _I don't
think any are necessary._
- [ ] My changes generate no new warnings - _I doubt it, but I have not
been able to install everything, I discovered this issue on my first
attempt to build._
- [ ] I have added tests that prove my fix is effective or that my
feature works - _Not a feature._
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file - _Not sure which file that is
in this case, please let me know if this is applicable._
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

---------

Co-authored-by: Kelly Guo <kellyg@nvidia.com>