Fixes MARL workflows for recording videos during training/inferencing (#1596)
# Description

Fixing bug so that using training workflow on MARL workflow populates
videos/train.
See #1595

## Type of change
- Bug fix (non-breaking change which fixes an issue)
## Screenshots

![before_and_after](https://github.com/user-attachments/assets/5b662a88-dedd-4220-a0c4-8e7d09ceb51f)
The first run was without the changes where we see videos/train empty.
The second run is after the changes with videos/train successfully
populated.

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [N/A] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [Sort of] I have added tests that prove my fix is effective or that my
feature works; I have verified that it works on train.py for skrl and
rl_games. I have not verified rsl_rl or sb3 as well have not verified
play.py on any of the four. However I have implemented the changes on
all of them as they all seem to follow the exact same structure.
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [ ] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there; Unsure if this fix is worth being labelled as a
contributor, if so would be happy to be added to the contributors.md
(full name is Rishi Veerapaneni).