Adds a tutorial for policy inference in a prebuilt USD scene (#231)
# Description

This PR adds a tutorial to show how to inference on a trained policy in
a prebuilt USD scene.
It includes a script that performs inference of the
`Isaac-Velocity-Rough-H1-v0` environment in a warehouse scene and an
accompanying tutorial that walks through the script implementation.


## Type of change

- This change requires a documentation update

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

<!--
As you go through the checklist above, you can mark something as done by
putting an x character in it

For example,
- [x] I have done this task
- [ ] I have not done this task
-->

---------

Co-authored-by: Kelly Guo <kellyguo123@hotmail.com>