Adds option to filter collisions and real-time playback (#253)
This change introduces a new setting in InteractiveSceneCfg to allow
specifying whether collision filtering across environments is desired.
Note that when using the direct workflow, this option will only be
applied automatically when replicate_physics is also enabled in the
config. If replicate_physics is not enabled, a warning will appear in
the logs to prompt users to make a call to scene.filter_collisions().
This is required because collision filtering happens as part of the
physics replication process.

In addition, real-time playback support is extended for all RL library
play.py scripts, allowing real-time replay when possible for
inferencing.

<!-- As you go through the list, delete the ones that are not
applicable. -->

- Bug fix (non-breaking change which fixes an issue)
- New feature (non-breaking change which adds functionality)
- This change requires a documentation update

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [ ] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

<!--
As you go through the checklist above, you can mark something as done by
putting an x character in it

For example,
- [x] I have done this task
- [ ] I have not done this task
-->