Adds interval resampling on event manager's reset call (#1750)
# Description

Previously on episodic resets, the "time left" for the interval events
were not getting resampled. This means that if the user expects the
event to happen in the range 6-8s of an episode, it could be that in a
new episode, the push happens at a time outside this range as the time
left variable keeps its old value.

This MR fixes this issue by resampling the time left inside the
manager's reset call.

Note: This only matters for the case when "is_global_time" is False
(which is the default). The reason is that when there is global time,
the events are triggered through simulation time and not episode time.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there