Configures XR teleop camera placement per-task with an XrCfg env configuration (#282)
Currently, the camera placement for the stacking task is set to a
task-specific
view in each main script (record_demos.py, teleop_se3_agent.py). This is
not
scalable for when we need to add more tasks with different camera
placements,
or more main scripts.

Instead, this change adds an XrCfg with "xr_anchor_pos" and
"xr_anchor_rot" fields
which can specify the XR anchor prim per-task (e.g. in the stacking base
task).

Specifically: the pose of this prim will be placed at the origin of the
XR
device's local coordinate frame. This is achieved by setting the XR
anchor mode
to "custom anchor", which determines the coordinate frame transformation
between
the simulation and the XR device's local frame.

The default camera placement is set to the scene origin pose (in case a
custom
anchor is not set for a task).

- Bug fix (non-breaking change which fixes an issue)

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there