Adds tutorial for training & validating HOVER policy using Isaac Lab (#320)
# Description

<!--
Thank you for your interest in sending a pull request. Please make sure
to check the contribution guidelines.

Link:
https://isaac-sim.github.io/IsaacLab/main/source/refs/contributing.html
-->

This MRs adds a tutorial for training and validating the Hover policy
(already released to public https://github.com/NVlabs/HOVER) in Isaac
Lab.

<!-- As a practice, it is recommended to open an issue to have
discussions on the proposed pull request.
This makes it easier for the community to keep track of what is being
developed or added, and if a given feature
is demanded by more than one party. -->

## Type of change

<!-- As you go through the list, delete the ones that are not
applicable. -->

- Added a tutorial to the documentation outlining the steps for training
and evaluation.

## Screenshots


![image](https://github.com/user-attachments/assets/44fff773-2484-499d-b3e9-8b6d54efc387)

![image](https://github.com/user-attachments/assets/478e6908-31e1-4873-b338-4182847992a7)

![image](https://github.com/user-attachments/assets/77056f8b-b4c5-4f97-be90-bb9bd75eb4c0)

![image](https://github.com/user-attachments/assets/97432a3c-6959-4b18-a2ac-cc9a39043d54)


<!--
Example:

| Before | After |
| ------ | ----- |
| _gif/png before_ | _gif/png after_ |

To upload images to a PR -- simply drag and drop an image while in edit
mode and it should upload the image directly. You can then paste that
source into the above before/after sections.
-->

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

<!--
As you go through the checklist above, you can mark something as done by
putting an x character in it

For example,
- [x] I have done this task
- [ ] I have not done this task
-->